requirements.txt
flask==3.0.3
supabase==2.4.6
python-dotenv==1.0.1
flask-login==0.6.3
Werkzeug==3.0.3
requests==2.31.0
reportlab==4.4.10
pillow==12.2.0
gunicorn==22.0.0
